Moves product to file
